feat: Migrated Data Module To KMP
