+++ core/data/src/androidMain/AndroidManifest.xml
