+++ core/data/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/data/model/NewsResource.kt
