+++ core/data/src/commonTest/kotlin/com/google/samples/apps/nowinandroid/core/data/repository/TestNotifier.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.data.repository
+
+import com.google.samples.apps.nowinandroid.core.model.data.NewsResource
+import com.google.samples.apps.nowinandroid.core.notifications.Notifier
+
+/**
+ * A test implementation of the [Notifier] interface that collects the posted news resources.
+ * This is useful for testing purposes to verify that the correct news resources are being posted.
+ * This class is not intended for production use.
+ * and could be moved into testing module when it's migrated to KMP
+ */
+class TestNotifier : Notifier {
+    val addedNewsResources = mutableListOf<List<NewsResource>>()
+
+    override fun postNewsNotifications(newsResources: List<NewsResource>) {
+        addedNewsResources.add(newsResources)
+    }
+}
