+++ core/data/src/commonTest/kotlin/com/google/samples/apps/nowinandroid/core/data/testdoubles/TestNiaNetworkDataSource.kt
@@ -1,0 +1,125 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.data.testdoubles
+
+import com.google.samples.apps.nowinandroid.core.data.util.runBlockingExt
+import com.google.samples.apps.nowinandroid.core.network.NiaNetworkDataSource
+import com.google.samples.apps.nowinandroid.core.network.demo.DemoNiaNetworkDataSource
+import com.google.samples.apps.nowinandroid.core.network.model.NetworkChangeList
+import com.google.samples.apps.nowinandroid.core.network.model.NetworkNewsResource
+import com.google.samples.apps.nowinandroid.core.network.model.NetworkTopic
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
+import kotlinx.serialization.json.Json
+
+enum class CollectionType {
+    Topics,
+    NewsResources,
+}
+
+/**
+ * Test double for [NiaNetworkDataSource]
+ */
+class TestNiaNetworkDataSource : NiaNetworkDataSource {
+
+    @OptIn(ExperimentalCoroutinesApi::class)
+    private val source = DemoNiaNetworkDataSource(
+        UnconfinedTestDispatcher(),
+        Json { ignoreUnknownKeys = true },
+    )
+
+    private val allTopics: List<NetworkTopic> = runBlockingExt { source.getTopics() }
+
+    private val allNewsResources: List<NetworkNewsResource> =
+        runBlockingExt { source.getNewsResources() }
+
+    private val changeLists: MutableMap<CollectionType, List<NetworkChangeList>> = mutableMapOf(
+        CollectionType.Topics to allTopics
+            .mapToChangeList(idGetter = NetworkTopic::id),
+        CollectionType.NewsResources to allNewsResources
+            .mapToChangeList(idGetter = NetworkNewsResource::id),
+    )
+
+    override suspend fun getTopics(ids: List<String>?): List<NetworkTopic> =
+        allTopics.matchIds(
+            ids = ids,
+            idGetter = NetworkTopic::id,
+        )
+
+    override suspend fun getNewsResources(ids: List<String>?): List<NetworkNewsResource> =
+        allNewsResources.matchIds(
+            ids = ids,
+            idGetter = NetworkNewsResource::id,
+        )
+
+    override suspend fun getTopicChangeList(after: Int?): List<NetworkChangeList> =
+        changeLists.getValue(CollectionType.Topics).after(after)
+
+    override suspend fun getNewsResourceChangeList(after: Int?): List<NetworkChangeList> =
+        changeLists.getValue(CollectionType.NewsResources).after(after)
+
+    fun latestChangeListVersion(collectionType: CollectionType) =
+        changeLists.getValue(collectionType).last().changeListVersion
+
+    fun changeListsAfter(collectionType: CollectionType, version: Int) =
+        changeLists.getValue(collectionType).after(version)
+
+    /**
+     * Edits the change list for the backing [collectionType] for the given [id] mimicking
+     * the server's change list registry
+     */
+    fun editCollection(collectionType: CollectionType, id: String, isDelete: Boolean) {
+        val changeList = changeLists.getValue(collectionType)
+        val latestVersion = changeList.lastOrNull()?.changeListVersion ?: 0
+        val change = NetworkChangeList(
+            id = id,
+            isDelete = isDelete,
+            changeListVersion = latestVersion + 1,
+        )
+        changeLists[collectionType] = changeList.filterNot { it.id == id } + change
+    }
+}
+
+fun List<NetworkChangeList>.after(version: Int?): List<NetworkChangeList> = when (version) {
+    null -> this
+    else -> filter { it.changeListVersion > version }
+}
+
+/**
+ * Return items from [this] whose id defined by [idGetter] is in [ids] if [ids] is not null
+ */
+private fun <T> List<T>.matchIds(
+    ids: List<String>?,
+    idGetter: (T) -> String,
+) = when (ids) {
+    null -> this
+    else -> ids.toSet().let { idSet -> filter { idGetter(it) in idSet } }
+}
+
+/**
+ * Maps items to a change list where the change list version is denoted by the index of each item.
+ * [after] simulates which models have changed by excluding items before it
+ */
+private fun <T> List<T>.mapToChangeList(
+    idGetter: (T) -> String,
+) = mapIndexed { index, item ->
+    NetworkChangeList(
+        id = idGetter(item),
+        changeListVersion = index + 1,
+        isDelete = false,
+    )
+}
